kansai/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
cloudinary
python-multipart
websockets
alembic
pydantic
pydantic-settings
